requirements.txt
pillow==4.0.0
numpy==1.11.3